Inventory Inclussion
After much hassle have added an inventory class that we can add items to
easily, currently switching between items works well, and I have changed
the way most of the scripts work for easier use in future. Slight
improvement to the spider AI so they are only scared of the player when
a gun is drawen. Only thing left to do is to manage a way to
differentiate between when the player dusts, and when the player tries
to interact.